pyproject.toml
[tool.poetry]
name = "duplicates-service"
version = "0.1.0"
description = "Service for detecting duplicate transactions"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
gunicorn = "^21.2.0"
redis = "^5.0.1"
google-cloud-bigquery = "^3.17.0"
google-cloud-pubsub = "^2.19.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.9.1"
pyyaml = "^6.0.2"
httpx = "^0.27.0"
pandas = "^2.2.3"
pandas-gbq = "^0.28.0"
levenshtein = "^0.27.1"
jellyfish = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 